Allows routes to be specified as 'http'.
Issue #2536 fixed an issue where URL::to() couldn't be used to create
http URLs from an https page.  This is a fix to a similar but different
issue where URL::route() can't be used to create routes to URLs that are
http if the current page is https.

In order to properly fix this issue with consistency for the current
route design, it was necessary to add an 'http' option to the route
actions (an 'https' option already existed).

So with this commit it is now possible to specify 'http' in a route
declaration, which behaves exactly as the 'https' option does, forcing
generated route URLs to use http, and only responding to http requests
for that route.